+++ src/state-transform.js
@@ -17,6 +17,10 @@ export function transformState(entityId, newState, hass, entities, i18n) {
   const cfg = getCustomConfig(entityId, entities);
   const rawState = newState.state;
   const unit = newState.attributes?.unit_of_measurement;
+  const entity_picture =
+    newState.attributes?.entity_picture ||
+    newState.attributes?.entity_picture_local ||
+    null;
 
   const name = getEntityName(entityId, entities, hass.states);
   const icon = getIconForEntity(hass.states[entityId], cfg, rawState);
@@ -37,6 +41,7 @@ export function transformState(entityId, newState, hass, entities, i18n) {
     name,
     icon,
     icon_color,
+    entity_picture,
     state: stateWithUnit,
     raw_state: rawState,
     time: new Date(newState.last_changed), // must be Date() for relative_time