+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## v1.6.0
+- Fixed: card_mod compatibility and localize empty state message by **@kvanzuijlen**, Thank you!
+- Added Swedish translations by **@naitkris**, Thank you!
+- Added card option `card_background`  to set background color
+- Added card options `timeline_color_start`, `timeline_color_end`, `dot_color` to set timeline & dots color
+- Added entity option `show_entity_picture` to show the entity picture instead of icon if available 
+- Added transparency slider to all color pickers
+
+
+
 ## v1.5.1
 - Resolved a CustomElementRegistry conflict with the LLM Vision Card by
   renaming the internal editor element to a unique identifier.